+++ docs/examples/tiledata/tooltip.js
@@ -4,7 +4,7 @@ import { Flex, TileData } from 'gestalt';
 
 export default function Example(): Node {
   return (
-    <Flex justifyContent="center">
+    <Flex justifyContent="center" alignItems="center" width="100%" height="100%">
       <TileData tooltip={{ text: 'Weekly Active Users' }} title="WAU" value="1.25M" selected />{' '}
     </Flex>
   );