+++ packages/gestalt/src/index.js
@@ -78,6 +78,7 @@ import ColorSchemeProvider, { useColorScheme } from './contexts/ColorSchemeProvi
 import DefaultLabelProvider from './contexts/DefaultLabelProvider.js';
 import DeviceTypeProvider from './contexts/DeviceTypeProvider.js';
 import ExperimentProvider from './contexts/ExperimentProvider.js';
+import GlobalEventsHandlerProvider from './contexts/GlobalEventsHandlerProvider.js';
 import OnLinkNavigationProvider from './contexts/OnLinkNavigationProvider.js';
 
 export {
@@ -105,6 +106,7 @@ export {
   Fieldset,
   FixedZIndex,
   Flex,
+  GlobalEventsHandlerProvider,
   Heading,
   HelpButton,
   Icon,