+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+## 112.5.1 (Jun 1, 2023)
+
+### Patch
+
+- Docs: Fix links on Android docs pages ([#2973](https://github.com/pinterest/gestalt/pull/2973)) - [Preview link](https://deploy-preview-2973--gestalt.netlify.app)
+
+## 112.5.0 (Jun 1, 2023)
+
+### Minor
+
+- Docs: Update Designers page ([#2971](https://github.com/pinterest/gestalt/pull/2971)) - [Preview link](https://deploy-preview-2971--gestalt.netlify.app)
+
+## 112.4.0 (Jun 1, 2023)
+
+### Minor
+
+- GlobalEventsHandlerProvider: new provider to share global event handlers to consuming components ([#2959](https://github.com/pinterest/gestalt/pull/2959)) - [Preview link](https://deploy-preview-2959--gestalt.netlify.app)
+
+## 112.3.9 (Jun 1, 2023)
+
+### Patch
+
+- Table: Convert examples to use Sandpack ([#2969](https://github.com/pinterest/gestalt/pull/2969)) - [Preview link](https://deploy-preview-2969--gestalt.netlify.app)
+
+## 112.3.8 (Jun 1, 2023)
+
+### Patch
+
+- Docs: Added Android specs for Avatar, Button, Checkbox, IconButton, Searchfield, Sheet, Switch, Tabs, Textfield ([#2972](https://github.com/pinterest/gestalt/pull/2972)) - [Preview link](https://deploy-preview-2972--gestalt.netlify.app)
+
+## 112.3.7 (May 30, 2023)
+
+### Patch
+
+- TileData: Add Typescript Types ([#2966](https://github.com/pinterest/gestalt/pull/2966)) - [Preview link](https://deploy-preview-2966--gestalt.netlify.app)
+
+## 112.3.6 (May 30, 2023)
+
+### Patch
+
+- SlimBanner: Convert examples to use Sandpack ([#2968](https://github.com/pinterest/gestalt/pull/2968)) - [Preview link](https://deploy-preview-2968--gestalt.netlify.app)
+
+## 112.3.5 (May 26, 2023)
+
+### Patch
+
+- Docs: Weekly update and roadmap updates ([#2967](https://github.com/pinterest/gestalt/pull/2967)) - [Preview link](https://deploy-preview-2967--gestalt.netlify.app)
+
+## 112.3.4 (May 26, 2023)
+
+### Patch
+
+- Callout: convert examples to use Sandpack ([#2960](https://github.com/pinterest/gestalt/pull/2960)) - [Preview link](https://deploy-preview-2960--gestalt.netlify.app)
+
+## 112.3.3 (May 26, 2023)
+
+### Patch
+
+- Internal: Fix TileData Examples ([#2965](https://github.com/pinterest/gestalt/pull/2965)) - [Preview link](https://deploy-preview-2965--gestalt.netlify.app)
+
+## 112.3.2 (May 26, 2023)
+
+### Patch
+
+- TileData: add accessibility pause ([#2964](https://github.com/pinterest/gestalt/pull/2964)) - [Preview link](https://deploy-preview-2964--gestalt.netlify.app)
+
+## 112.3.1 (May 25, 2023)
+
+### Patch
+
+- Internal: Flow fixes ([#2961](https://github.com/pinterest/gestalt/pull/2961)) - [Preview link](https://deploy-preview-2961--gestalt.netlify.app)
+
 ## 112.3.0 (May 24, 2023)
 
 ### Minor